Convert ESRI viewers to controller pattern
