+++ app/assets/stylesheets/geoblacklight/modules/index_maps.scss
@@ -6,7 +6,7 @@
   }
 }
 
-.viewer[data-protocol="IndexMap"] {
+.viewer[data-leaflet-viewer-protocol-value="IndexMap"] {
   .leaflet-interactive {
     stroke: black;
     fill-opacity: 0.65;
