+++ app/javascript/leaflet/viewers/esri/tiled_map_layer.js
@@ -2,29 +2,8 @@ import { tiledMapLayer, identifyFeatures } from "esri-leaflet";
 import LeafletViewerEsri from "../esri.js";
 
 export default class LeafletViewerEsriTiledMapLayer extends LeafletViewerEsri {
-  getPreviewLayer() {
-    // Set layer URL
-    this.options.url = this.data.url;
-
-    // Check if this is a tile map layer and for correct spatial reference
-    if (
-      this.layerInfo.singleFusedMapCache === true &&
-      this.layerInfo.spatialReference.wkid === 102100
-    ) {
-      /**
-       * TODO: allow non-mercator projections and custom scales
-       *       - use Proj4Leaflet
-       */
-
-      const esriTiledMapLayer = tiledMapLayer(this.options);
-
-      // Setup feature inspection
-      this.setupInspection(esriTiledMapLayer);
-
-      return esriTiledMapLayer;
-    }
-  }
-
+  // TODO: This method needs to be ported over to a new inspection.js file
+  // can delete this file when done.
   setupInspection(layer) {
     this.map.on("click", (e) => {
       this.appendLoadingMessage();
